replace management command install_workdir by initialize_shop_demo
